refactor user header component
