+++ src/app/[auth]/explore/page.tsx
@@ -1,20 +1,17 @@
-/* eslint-disable @next/next/no-img-element */
 'use client';
 
 import { getUsersAll, getUsersBySearch } from '@/api/user';
-import UserHeader from '@/components/Header/userHeader';
 import ToggleAvailable from '@/components/common/near-wallet/toggleAvailable';
 import SearchByLocation from '@/components/explore/searchByLocation';
 import SearchBySkills from '@/components/explore/searchBySkills';
 import UserOverView from '@/components/explore/userOverview';
+import { UserHeader } from '@/components/header/';
 import { TOTAL_SKILLS } from '@/constant/skills';
 import { useAppDispatch, useAppSelector } from '@/redux/hook';
 import { setUsers } from '@/redux/userSlice';
 import UserType from '@/types/user';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
-
-/* eslint-disable @next/next/no-img-element */
 
 const Explore = () => {
   const router = useRouter();
