+++ src/app/[auth]/layout.tsx
@@ -1,24 +1,18 @@
 'use client';
 
+import { Header } from '@/components/header/header';
+import Sidebar from '@/components/sidebar';
+import { useAppSelector } from '@/redux/hook';
 import React from 'react';
 
-import Header from '@/components/Header/header';
-import Sidebar from '@/components/sidebar';
-
-import { useAppSelector } from '@/redux/hook';
-
 const Layout = ({ children }: { children: React.ReactNode }) => {
-  
   const user = useAppSelector((state) => state.userReducer);
 
   return (
     <>
       <Header />
       {children}
-      <Sidebar
-        openSideBar={user.toggleOpenSidebar}
-        userInfo={user.user}
-      />
+      <Sidebar openSideBar={user.toggleOpenSidebar} userInfo={user.user} />
     </>
   );
 };
