+++ src/app/[auth]/notification/page.tsx
@@ -1,9 +1,7 @@
-/* eslint-disable @next/next/no-img-element */
 'use client';
 
+import { Header } from '@/components/header';
 import React from 'react';
-
-import Header from '@/components/Header/userHeader';
 
 export default function Notification() {
   return (
