+++ src/app/page.tsx
@@ -1,7 +1,17 @@
 'use client';
 
+import { Header } from '@/components/header';
+import { Profile } from '@/components/profile';
+
 const RedirectScreen = () => {
-  return <div>Welcome to home page!</div>;
+  return (
+    <>
+      <Header />
+      <div className='container mx-auto py-4'>
+        <Profile />
+      </div>
+    </>
+  );
 };
 
 export default RedirectScreen;
