+++ src/components/Header/index.ts
@@ -1,0 +1,5 @@
+import { ExploreHeader } from "./explore-header";
+import { Header } from "./header";
+import { UserHeader } from "./user-header";
+
+export { ExploreHeader, Header, UserHeader };
