+++ src/components/profile/index.ts
@@ -1,0 +1,3 @@
+import { Profile } from "./profile";
+
+export { Profile };
