+++ src/components/profile/profile.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const Profile: React.FC = () => {
+  return (
+    <div>
+      <div>meow meow</div>
+    </div>
+  );
+};
